# micro-optimisations: cached hashcode and reduce hashmap lookups