--- a/src/main/scala/ir/Expr.scala
+++ b/src/main/scala/ir/Expr.scala
@@ -1,5 +1,6 @@
 package ir
 import boogie._
+import util.CachedHashCode
 import scala.collection.mutable
 
 sealed trait Expr {
@@ -45,14 +46,14 @@ case object FalseLiteral extends BoolLit {
   override def value = false
 }
 
-case class BitVecLiteral(value: BigInt, size: Int) extends Literal {
+case class BitVecLiteral(value: BigInt, size: Int) extends Literal with CachedHashCode {
   assert(size >= 0)
   override def toBoogie: BitVecBLiteral = BitVecBLiteral(value, size)
   override def getType: IRType = BitVecType(size)
   override def toString: String = s"${value}bv$size"
 }
 
-case class IntLiteral(value: BigInt) extends Literal {
+case class IntLiteral(value: BigInt) extends Literal with CachedHashCode {
   override def toBoogie: IntBLiteral = IntBLiteral(value)
   override def getType: IRType = IntType
   override def toString: String = value.toString
@@ -68,7 +69,7 @@ case class IntLiteral(value: BigInt) extends Literal {
   *
   * Requires end > start
   */
-case class Extract(end: Int, start: Int, body: Expr) extends Expr {
+case class Extract(end: Int, start: Int, body: Expr) extends Expr with CachedHashCode {
   override def toBoogie: BExpr = BVExtract(end, start, body.toBoogie)
   override def gammas: Set[Variable] = body.gammas
   override def variables: Set[Variable] = body.variables
@@ -81,7 +82,7 @@ case class Extract(end: Int, start: Int, body: Expr) extends Expr {
   *
   * Requires repeats > 0.
   */
-case class Repeat(repeats: Int, body: Expr) extends Expr {
+case class Repeat(repeats: Int, body: Expr) extends Expr with CachedHashCode {
   override def toBoogie: BExpr = BVRepeat(repeats, body.toBoogie)
   override def gammas: Set[Variable] = body.gammas
   override def variables: Set[Variable] = body.variables
@@ -95,7 +96,7 @@ case class Repeat(repeats: Int, body: Expr) extends Expr {
 }
 
 /** Zero-extends by extension extra bits. */
-case class ZeroExtend(extension: Int, body: Expr) extends Expr {
+case class ZeroExtend(extension: Int, body: Expr) extends Expr with CachedHashCode {
   override def toBoogie: BExpr = BVZeroExtend(extension, body.toBoogie)
   override def gammas: Set[Variable] = body.gammas
   override def variables: Set[Variable] = body.variables
@@ -109,7 +110,7 @@ case class ZeroExtend(extension: Int, body: Expr) extends Expr {
 }
 
 /** Sign-extends by extension extra bits. */
-case class SignExtend(extension: Int, body: Expr) extends Expr {
+case class SignExtend(extension: Int, body: Expr) extends Expr with CachedHashCode {
   override def toBoogie: BExpr = BVSignExtend(extension, body.toBoogie)
   override def gammas: Set[Variable] = body.gammas
   override def variables: Set[Variable] = body.variables
@@ -122,7 +123,7 @@ case class SignExtend(extension: Int, body: Expr) extends Expr {
   override def acceptVisit(visitor: Visitor): Expr = visitor.visitSignExtend(this)
 }
 
-case class UnaryExpr(op: UnOp, arg: Expr) extends Expr {
+case class UnaryExpr(op: UnOp, arg: Expr) extends Expr with CachedHashCode {
   override def toBoogie: BExpr = UnaryBExpr(op, arg.toBoogie)
   override def gammas: Set[Variable] = arg.gammas
   override def variables: Set[Variable] = arg.variables
@@ -171,7 +172,7 @@ sealed trait BVUnOp(op: String) extends UnOp {
 case object BVNOT extends BVUnOp("not")
 case object BVNEG extends BVUnOp("neg")
 
-case class BinaryExpr(op: BinOp, arg1: Expr, arg2: Expr) extends Expr {
+case class BinaryExpr(op: BinOp, arg1: Expr, arg2: Expr) extends Expr with CachedHashCode {
   override def toBoogie: BExpr = BinaryBExpr(op, arg1.toBoogie, arg2.toBoogie)
   override def gammas: Set[Variable] = arg1.gammas ++ arg2.gammas
   override def variables: Set[Variable] = arg1.variables ++ arg2.variables
@@ -321,7 +322,7 @@ enum Endian {
   case BigEndian
 }
 
-case class UninterpretedFunction(name: String, params: Seq[Expr], returnType: IRType) extends Expr {
+case class UninterpretedFunction(name: String, params: Seq[Expr], returnType: IRType) extends Expr with CachedHashCode {
   override def getType: IRType = returnType
   override def toBoogie: BFunctionCall = BFunctionCall(name, params.map(_.toBoogie).toList, returnType.toBoogie, true)
   override def acceptVisit(visitor: Visitor): Expr = visitor.visitUninterpretedFunction(this)
@@ -361,7 +362,7 @@ object Variable {
   * These are variables with global scope (in a 'accessible from any procedure' sense), not related to the concurrent
   * shared memory sense.
   */
-case class Register(override val name: String, size: Int) extends Variable with Global {
+case class Register(override val name: String, size: Int) extends Variable with Global with CachedHashCode {
   override def toGamma: BVar = BVariable(s"Gamma_$name", BoolBType, Scope.Global)
   override def toBoogie: BVar = BVariable(s"$name", irType.toBoogie, Scope.Global)
   override def toString: String = s"Register(${name}, $irType)"
@@ -370,7 +371,9 @@ case class Register(override val name: String, size: Int) extends Variable with 
 }
 
 /** Variable with scope local to the procedure, typically a temporary variable created in the lifting process. */
-case class LocalVar(varName: String, override val irType: IRType, val index: Int = 0) extends Variable {
+case class LocalVar(varName: String, override val irType: IRType, val index: Int = 0)
+    extends Variable
+    with CachedHashCode {
   override val name = varName + (if (index > 0) then s"_$index" else "")
   override def toGamma: BVar = BVariable(s"Gamma_$name", BoolBType, Scope.Local)
   override def toBoogie: BVar = BVariable(s"$name", irType.toBoogie, Scope.Local)